requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
rdflib==7.0.0
python-dotenv==1.0.0

# LLM Providers
openai==1.3.0
anthropic==0.8.0
google-generativeai==0.3.0

# Other ML/NLP
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.12
sentence-transformers==2.2.2
faiss-cpu==1.7.4
tiktoken==0.5.2

# Web
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1

# Utils
httpx==0.25.0
lxml==4.9.3
tqdm==4.66.0
